Refine usage of hpcrun_safe_enter/exit
hpcrun_safe_enter/exit provides a way for code to mark itself as "inside Run," for
the purposes of asynchronous interrupts. Samples taken "inside Run" are dropped or
otherwise to prevent a number of issues.

The documentation for these functions is a bit rough, but from code inspection
the "correct" usage should support nested cases as well. Even so, a number of
code locations don't follow this "correct" usage. These seem to be callbacks from
"outside," but 1) this makes the usage inconsistent and 2) they may be nested due
to recursive calls through external libraries (we've fought off enough of them lately).

This patch adds an attribute to hpcrun_safe_enter to warn when its return value is
discarded, and then cleans up the resulting warnings by applying the nested
instructions to all current usage. No functional change.