Rewrite ParallelForEach to set the action during fill
There's no particular reason why the action needs to be constant throughout the
entire lifetime of the object, and setting it later allows for data to be saved
in the callable itself for the duration of the loop.

Some bits are still up for debate, but this works well enough for now.