Remove the dependence on external Valgrind headers for Valgrind annotations
Valgrind isn't always distributed with its headers in a reasonable place, so
instead of relying on external headers copy the ones we need into the codebase.
The Valgrind client headers are intentionally under a weaker license than the
rest of the code for this exact purpose.

Rather than using a #define in hpctoolkit-config.h, now configure will define
NVALGRIND if the annotations were not requested. This is more-or-less
equivalent to NDEBUG for assert, so its easier to control.

The Helgrind header has a few extra additions to be warning-free in the
NVALGRIND case, this is noted in the file's header as per the legalese.